Fixing problem with color tag at start of string
