+++ CustomCommands/Services/PluginUtilities.cs
@@ -57,108 +57,41 @@ public class PluginUtilities : IPluginUtilities
             return true;
         }
     }
-
     /// <summary>
-    /// Checks if the command is on cooldown
+    /// Moves the color tag one space the the left if it's not already there.
+    /// Because the game doesn't support color tags at the start of a string.
     /// </summary>
-    /// <param name="player"></param>
-    /// <param name="cmd"></param>
+    /// <param name="input"></param>
     /// <returns></returns>
-    public bool IsCommandOnCooldown(CCSPlayerController player, Commands cmd)
+    public string PadLeftColorTag(string input)
     {
-        // Check global cooldown
-        if (IsCommandOnCooldownWithCondition(x => x.IsGlobal == true && x.CommandID == cmd.ID, player, cmd))
-            return true;
-
-        // Check player cooldown
-        if (IsCommandOnCooldownWithCondition(x => x.PlayerID == player.UserId && x.CommandID == cmd.ID, player, cmd))
-            return true;
-
-        return false;
-    }
-
-    private bool IsCommandOnCooldownWithCondition(Func<CooldownTimer, bool> predicate, CCSPlayerController player, Commands cmd)
-    {
-        int index = PluginGlobals.CooldownTimer.FindIndex(x => predicate(x) && x.CooldownTime > DateTime.Now);
-
-        if (index != -1)
-        {
-            string timeleft = PluginGlobals.CooldownTimer[index].CooldownTime.Subtract(DateTime.Now).Seconds.ToString();
-            player.PrintToChat($"{PluginGlobals.Config.Prefix}{cmd.Cooldown.CooldownMessage.Replace("{TIME}", timeleft) 
-                ?? $"This command is for {timeleft} seconds on cooldown"}");
-
-            return true;
-        }
-
-        return false;
-    }
-    
-    /// <summary>
-    /// Adds the command to the cooldown list
-    /// </summary>
-    /// <param name="isGlobal"></param>
-    /// <param name="playerID"></param>
-    /// <param name="commandID"></param>
-    /// <param name="cooldownTime"></param>
-    public void AddToCooldownList(bool isGlobal, int playerID, Guid commandID, int cooldownTime)
-    {
-        var timer = new CooldownTimer() {
-            IsGlobal = isGlobal, 
-            CommandID = commandID, 
-            CooldownTime = DateTime.Now.AddSeconds(cooldownTime)
+        string[] colorTagList = new string[] {
+            "{DEFAULT}",
+            "{WHITE}",
+            "{DARKRED}",
+            "{RED}",
+            "{LIGHTRED}",
+            "{GREEN}",
+            "{LIME}",
+            "{OLIVE}",
+            "{ORANGE}",
+            "{GOLD}",
+            "{YELLOW}",
+            "{BLUE}",
+            "{DARKBLUE}",
+            "{LIGHTPURPLE}",
+            "{PURPLE}",
+            "{SILVER}",
+            "{BLUEGREY}",
+            "{GREY}",
         };
-
-        if (isGlobal)
+        foreach (var colorTag in colorTagList)
         {
-            int index = PluginGlobals.CooldownTimer.FindIndex(x => 
-                x.IsGlobal == true 
-                && x.CommandID == commandID);
-            if (index != -1)
-                PluginGlobals.CooldownTimer[index].CooldownTime = timer.CooldownTime;
-            else
-                PluginGlobals.CooldownTimer.Add(timer);
-        }
-        else
-        {
-            timer.PlayerID = playerID;
-            int index = PluginGlobals.CooldownTimer.FindIndex(x => 
-                x.PlayerID == playerID 
-                && x.CommandID == commandID);
-            if (index != -1)
-                PluginGlobals.CooldownTimer[index].CooldownTime = timer.CooldownTime;
-            else
-                PluginGlobals.CooldownTimer.Add(timer);
-        }
-    }
-
-    /// <summary>
-    /// Sets the cooldown for the command
-    /// </summary>
-    /// <param name="player">Need to add the player if the Cooldown is only for a specific player</param>
-    /// <param name="cmd"></param>
-    public void SetCooldown(CCSPlayerController player, Commands cmd)
-    {
-        if (cmd.Cooldown is JsonElement jsonElement)
-        {
-            switch (jsonElement.ValueKind)
+            if (input.StartsWith(colorTag))
             {
-                case JsonValueKind.Number:
-                    int cooldown = (int)cmd.Cooldown;
-                    if (cooldown == 0) 
-                        break;
-
-                    AddToCooldownList(false, player.UserId ?? 0, cmd.ID, cooldown);
-                    break;
-
-                case JsonValueKind.Object:
-                    Cooldown cooldownObject = (Cooldown)cmd.Cooldown;
-
-                    AddToCooldownList(cooldownObject.IsGlobal, player.UserId ?? 0, cmd.ID, cooldownObject.CooldownTime);
-                    break;
-
-                default:
-                    break;
+                return " " + input;
             }
         }
+        return input;
     }
 }
