+++ CustomCommands/test.md
@@ -4,3 +4,4 @@
 * [ ] check if USERID tag is the right one from the console
 * [ ] check if the string pattern check works
 * [ ] check cooldown funktion for normal int and than object
+* [ ] check if padleft color tag works now
